[fix] lstick/rstick bounds
Issue: lsticks and rsticks used a hack to get alignment to the right / left which was based on taking double the width and placing the content on the far left / right respectively. For very short circuits, the bounding box of this phantom space could extend beyond the actual circuit.

Solution: We now make use of the feature of specifying an offset in the size-hint function.