[fix] equation numbering with circuits
When equation numbering is turned on, display-style braces as used with lstick, rstick and gate inputs/outputs, are not drawn correctly.

We could use non-block math but then the bounding box of the rectangles is not measured correctly. Instead we just turn equation numbering off for the entire circuit.

fixes #3