[change] replace default `fill: none` with `auto`
This is a breaking interface change and applies to `gate`, `mqgate`, `meter`, `targ`, `ctrl`, and `phase`.
The default behaviour has not changed. However, `auto` semantically fits this behaviour better since the default fill color is used as opposed to no fill. This change is suitable, since gates can now actually be transparent (due to wires not drawing through gates anymore).